requirements.txt
attrs==23.2.0
black==24.4.0
cattrs==23.2.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
flake8==7.0.0
fusepy==3.0.1
idna==3.7
jedi==0.19.1
mccabe==0.7.0
mypy==1.9.0
mypy-extensions==1.0.0
packaging==24.0
parso==0.8.4
pathspec==0.12.1
platformdirs==4.2.0
pycodestyle==2.11.1
pyflakes==3.2.0
python-dotenv==1.0.1
requests==2.31.0
requests-cache==1.2.0
six==1.16.0
tabulate==0.9.0
tomli==2.0.1
typing_extensions==4.11.0
url-normalize==1.4.3
urllib3==2.2.1
